refactor(core): split password.go to distinct files
