+++ pkg/password/compare.go
@@ -1,58 +1,12 @@
 package password
 
 import (
-	"crypto/rand"
 	"crypto/subtle"
 	"encoding/base64"
 	"fmt"
 	"golang.org/x/crypto/argon2"
 	"strings"
-
-	"chatty/chatty/app/config"
 )
-
-type params struct {
-	secret      []byte
-	memory      uint32
-	iterations  uint32
-	saltLength  uint32
-	keyLength   uint32
-	parallelism uint8
-}
-
-type Service struct {
-	params params
-}
-
-func NewService(cfg config.Password) Passworder {
-	return &Service{params: params{
-		secret:      []byte(cfg.Secret),
-		memory:      cfg.Memory,
-		iterations:  cfg.Iterations,
-		parallelism: cfg.Parallelism,
-		saltLength:  cfg.SaltLength,
-		keyLength:   cfg.KeyLength,
-	}}
-}
-
-func (e Service) Generate(password string) (string, error) {
-	salt, err := generateRandomBytes(e.params.saltLength)
-	if err != nil {
-		return "", err
-	}
-	secretSalt := append(salt, e.params.secret...)
-
-	hash := argon2.IDKey([]byte(password), secretSalt,
-		e.params.iterations, e.params.memory, e.params.parallelism, e.params.keyLength)
-
-	b64Salt := base64.RawStdEncoding.EncodeToString(salt)
-	b64Hash := base64.RawStdEncoding.EncodeToString(hash)
-
-	encodedHash := fmt.Sprintf("$argon2id$v=%d$m=%d,t=%d,p=%d,k=%d$%s$%s",
-		argon2.Version, e.params.memory, e.params.iterations, e.params.parallelism, e.params.keyLength, b64Salt, b64Hash)
-
-	return encodedHash, nil
-}
 
 func (e Service) Compare(password, encodedHash string) (bool, error) {
 	p, salt, hash, err := decodeHash(encodedHash)
@@ -67,16 +21,6 @@ func (e Service) Compare(password, encodedHash string) (bool, error) {
 	}
 
 	return false, nil
-}
-
-func generateRandomBytes(n uint32) ([]byte, error) {
-	b := make([]byte, n)
-	_, err := rand.Read(b)
-	if err != nil {
-		return nil, err
-	}
-
-	return b, nil
 }
 
 func decodeHash(encodedHash string) (p *params, salt, hash []byte, err error) {
