+++ pkg/password/errors.go
@@ -2,5 +2,5 @@ package password
 
 import "github.com/pkg/errors"
 
-var ErrInvalidHash = errors.New("the encoded hash is not in the correct format")
-var ErrIncompatibleVersion = errors.New("incompatible version of argon2")
+var ErrInvalidHash = errors.New("The encoded hash is not in the correct format")
+var ErrIncompatibleVersion = errors.New("Incompatible version of argon2")
